code/requirement.txt
tensorflow-gpu==1.12.0
h5py==2.8.0
pickleshare==0.7.5
pymesh2==0.2.1
scipy==1.3.0
scikit-learn==0.21.2
